Fixed Power Supply Compensation Networks
Changed the nets for the compensation networks for the TPS62873-Q1-based supplies to follow the guidance in the datasheet. Testing of the out-of-spec configuration indicates acceptable performance of 20-40mVpp noise independent of loading. Transient/step responses are also acceptable on all supplies except for the 1.2V supply, which experiences 90mVpk over/undershoots and 25mv output sag at 1A.